Add a way to add features manually.
MapQuery now supports layer.features() which works similar
to map.layers(). You can add a new feature with passing in
either a single or an array of objects. Those objects have
the format of:
{
    format: 'GeoJSON',
    geometry: {type: 'Point', coordinates: [34.2, 20.4]},
    properties: {foo: 'bar'}
}
The format defaults to 'GeoJSON'. Another format is KML, where
you put in a string of KML as a geometry.